auth api

+++ src/models/products/model.js
@@ -3,7 +3,7 @@
 const productsModel = (sequelize, DataTypes) => sequelize.define('Product', {
   name: { type: DataTypes.STRING, required: true },
   categories: { type: DataTypes.STRING, required: true },
-  price: { type: DataTypes.DECIMAL, required: true },
+  price: { type: DataTypes.FLOAT, required: true },
   description: {type: DataTypes.STRING, required:true},
   inStock: {type: DataTypes.INTEGER}
 });
